fixedUnit3HomeWork

+++ Unit1/src/Unit3/Unit3Task8.java
@@ -5,8 +5,8 @@ import java.util.Random;
 public class Unit3Task8 {
     public static void main(String[] args) {
         Random rnd = new Random();
-        double mass1[] = new double[10];
-        double mass2[] = new double[10];
+        int mass1[] = new int[10];
+        int mass2[] = new int[10];
         double mass3[] = new double[10];
         for (int i = 0; i < mass1.length; i++) {
             mass1[i] = rnd.nextInt(9);
@@ -21,7 +21,7 @@ public class Unit3Task8 {
         int count = 0;
         for (int i = 0; i < mass3.length; i++) {
             if (mass2[i] != 0) {
-                mass3[i] = mass1[i] / mass2[i];
+                mass3[i] =(double) mass1[i] / mass2[i];
             } else {
                 mass3[i] = 0;
             }
